third_party/blink/web_tests/external/wpt/css/css-align/content-distribution/place-content-shorthand-006-expected.txt

This is a testharness.js-based test.
Found 12 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Checking place-content: last baseline normal
  assert_equals: last baseline normal specified value expected "last baseline normal" but got ""
[FAIL] Checking place-content: last baseline left
  assert_equals: last baseline left specified value expected "last baseline left" but got ""
[FAIL] Checking place-content: last baseline right
  assert_equals: last baseline right specified value expected "last baseline right" but got ""
[FAIL] Checking place-content: last baseline start
  assert_equals: last baseline start specified value expected "last baseline start" but got ""
[FAIL] Checking place-content: last baseline end
  assert_equals: last baseline end specified value expected "last baseline end" but got ""
[FAIL] Checking place-content: last baseline center
  assert_equals: last baseline center specified value expected "last baseline center" but got ""
[FAIL] Checking place-content: last baseline flex-start
  assert_equals: last baseline flex-start specified value expected "last baseline flex-start" but got ""
[FAIL] Checking place-content: last baseline flex-end
  assert_equals: last baseline flex-end specified value expected "last baseline flex-end" but got ""
[FAIL] Checking place-content: last baseline stretch
  assert_equals: last baseline stretch specified value expected "last baseline stretch" but got ""
[FAIL] Checking place-content: last baseline space-around
  assert_equals: last baseline space-around specified value expected "last baseline space-around" but got ""
[FAIL] Checking place-content: last baseline space-between
  assert_equals: last baseline space-between specified value expected "last baseline space-between" but got ""
[FAIL] Checking place-content: last baseline space-evenly
  assert_equals: last baseline space-evenly specified value expected "last baseline space-evenly" but got ""
Harness: the test ran to completion.
